+++ Mods ChoGGi/RC Miner/Code/Script.lua
@@ -9,7 +9,7 @@ function OnMsg.ModsReloaded()
 	fire_once = true
 
 	-- version to version check with
-	local min_version = 34
+	local min_version = 36
 	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
 
 	-- if we can't find mod or mod is less then min_version (we skip steam since it updates automatically)
@@ -250,7 +250,7 @@ function PortableMiner:Idle(delay)
     self:ShowNotWorkingSign(false)
   end
 
-	Sleep(delay or 2500)
+	Sleep(type(delay) == "number" or 2500)
 
 	self:Gossip("Idle")
 	self:SetState("idle")